--- client/components/CreateAvatar/index.jsx
@@ -1,3 +1,0 @@
-import CreateAvatar from "./CreateAvatar";
-
-export default CreateAvatar;
