--- client/components/GenUsername/index.jsx
@@ -1,3 +1,0 @@
-import GenUsername from "./GenUsername";
-
-export default GenUsername;
