--- client/components/RegistrationCompleted/index.jsx
@@ -1,3 +1,0 @@
-import RegistrationCompleted from "./RegistrationCompleted";
-
-export default RegistrationCompleted;
