+++ client/components/SignUp/CreateAvatar.jsx
