+++ client/components/SignUp/RegistrationCompleted.jsx
