--- client/components/SignUpInfo/index.jsx
@@ -1,3 +1,0 @@
-import SignUpInfo from "./SignUpInfo";
-
-export default SignUpInfo;
