+++ client/components/SignUpProgressBar/SignUpProgressBar.scss
