--- client/components/SignUpProgressBar/index.jsx
@@ -1,3 +1,0 @@
-import SignUpProgressBar from "./SignUpProgressBar";
-
-export default SignUpProgressBar;
