--- client/components/TellUsAbout/index.jsx
@@ -1,3 +1,0 @@
-import TellUsAbout from "./TellUsAbout";
-
-export default TellUsAbout;
